wayland: Add support for fractional scaling
Example client: https://github.com/mahkoh/alacritty/tree/fractional-scaling

+++ src/platform_impl/linux/mod.rs
@@ -441,7 +441,7 @@ impl Window {
 
     #[inline]
     pub fn scale_factor(&self) -> f64 {
-        x11_or_wayland!(match self; Window(w) => w.scale_factor() as f64)
+        x11_or_wayland!(match self; Window(w) => w.scale_factor())
     }
 
     #[inline]
